+++ react/components/Autocomplete/components/TileList/TileList.tsx
@@ -16,6 +16,8 @@ interface TileListProps {
   totalProducts: number;
   layout: ProductLayout;
   isLoading: boolean;
+  onProductClick: (product: string, position: number) => void;
+  onSeeAllClick: (term: string) => void;
 }
 
 export class TileList extends React.Component<TileListProps> {
@@ -46,7 +48,7 @@ export class TileList extends React.Component<TileListProps> {
                     : "row",
               }}
             >
-              {this.props.products.map(product => {
+              {this.props.products.map((product, index: number) => {
                 const productSummary = ProductSummary.mapCatalogProductToProductSummary(
                   product,
                 );
@@ -54,11 +56,25 @@ export class TileList extends React.Component<TileListProps> {
                 return (
                   <li key={product.productId} className={styles.tileListItem}>
                     {this.props.layout === ProductLayout.Horizontal ? (
-                      <CustomListItem product={productSummary} />
+                      <CustomListItem
+                        product={productSummary}
+                        onClick={() => {
+                          this.props.onProductClick(
+                            productSummary.productId,
+                            index,
+                          );
+                        }}
+                      />
                     ) : (
                       <ExtensionPoint
                         id="product-summary"
                         product={productSummary}
+                        actionOnClick={() => {
+                          this.props.onProductClick(
+                            productSummary.productId,
+                            index,
+                          );
+                        }}
                       />
                     )}
                   </li>
@@ -72,6 +88,7 @@ export class TileList extends React.Component<TileListProps> {
                   to={`/${this.props.term}`}
                   query={`map=ft`}
                   className={styles.tileListSeeMore}
+                  onClick={() => this.props.onSeeAllClick(this.props.term)}
                 >
                   <FormattedMessage
                     id="store/seeMore"