+++ react/components/Autocomplete/index.tsx
@@ -14,6 +14,15 @@ import { IconClose, IconClock } from "vtex.styleguide";
 import { ProductListContext } from "vtex.product-list-context";
 import { withDevice } from "vtex.device-detector";
 import debounce from "debounce";
+import { withPixel } from "vtex.pixel-manager/PixelContext";
+import { withRuntime } from "../../utils/withRuntime";
+import {
+  EventType,
+  handleAutocompleteSearch,
+  handleItemClick,
+  handleProductClick,
+  handleSeeAllClick,
+} from "../../utils/pixel";
 
 const MAX_TOP_SEARCHES_DEFAULT = 10;
 const MAX_SUGGESTED_TERMS_DEFAULT = 5;
@@ -27,7 +36,7 @@ export enum ProductLayout {
 
 interface AutoCompleteProps {
   isOpen: boolean;
-  runtime: { account: string };
+  runtime: { page: string };
   inputValue: string;
   maxTopSearches: number;
   maxSuggestedTerms: number;
@@ -55,6 +64,7 @@ interface AutoCompleteProps {
   __unstableProductOrigin: "BIGGY" | "VTEX";
   __unstableProductOriginVtex: boolean;
   simulationBehavior: "default" | "skip" | null;
+  push: (data: any) => void;
 }
 
 interface AutoCompleteState {
@@ -156,6 +166,7 @@ class AutoComplete extends React.Component<
 
       this.setState({
         dynamicTerm: inputValue,
+        queryFromHover: undefined,
       });
 
       if (inputValue === null || inputValue === "") {
@@ -265,6 +276,18 @@ class AutoComplete extends React.Component<
       simulationBehavior,
     );
 
+    if (!queryFromHover) {
+      const { count, operator, misspelled } = result.data.productSuggestions;
+
+      handleAutocompleteSearch(
+        this.props.push,
+        operator,
+        misspelled,
+        count,
+        term,
+      );
+    }
+
     this.setState({
       isProductsLoading: false,
     });
@@ -363,6 +386,11 @@ class AutoComplete extends React.Component<
         showTitle={!hasSuggestion || !this.props.hideTitles}
         onItemHover={this.updateQueryByItemHover.bind(this)}
         showTitleOnEmpty={this.props.maxSuggestedTerms !== 0}
+        onItemClick={handleItemClick(
+          this.props.push,
+          this.props.runtime.page,
+          EventType.SearchSuggestionClick,
+        )}
       />
     );
   }
@@ -383,6 +411,11 @@ class AutoComplete extends React.Component<
             title={<FormattedMessage id={"store/topSearches"} />}
             items={this.state.topSearchedItems || []}
             showTitle={!this.props.hideTitles}
+            onItemClick={handleItemClick(
+              this.props.push,
+              this.props.runtime.page,
+              EventType.TopSearchClick,
+            )}
           />
         ) : null}
 
@@ -393,6 +426,11 @@ class AutoComplete extends React.Component<
             title={<FormattedMessage id={"store/history"} />}
             items={this.state.history || []}
             showTitle={!this.props.hideTitles}
+            onItemClick={handleItemClick(
+              this.props.push,
+              this.props.runtime.page,
+              EventType.HistoryClick,
+            )}
           />
         ) : null}
       </div>
@@ -417,6 +455,14 @@ class AutoComplete extends React.Component<
           totalProducts={this.state.totalProducts || 0}
           layout={this.getProductLayout()}
           isLoading={this.state.isProductsLoading}
+          onProductClick={handleProductClick(
+            this.props.push,
+            this.props.runtime.page,
+          )}
+          onSeeAllClick={handleSeeAllClick(
+            this.props.push,
+            this.props.runtime.page,
+          )}
         />
       </>
     );
@@ -522,4 +568,4 @@ class AutoComplete extends React.Component<
   }
 }
 
-export default withDevice(withApollo(AutoComplete));
+export default withPixel(withDevice(withApollo(withRuntime(AutoComplete))));