+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Pixel events to the autocomplete.
+
 ## [2.0.3] - 2020-09-17
 
 ### Fixed